fix percent calculation , should close #4252

+++ lib/libpacman/add.c
@@ -444,7 +444,7 @@ int _pacman_add_commit(pmtrans_t *trans, pmdb_t *db)
 				STRNCPY(pathname, archive_entry_pathname (entry), PATH_MAX);
 
 				if (info->size != 0)
-		    			percent = archive_position_uncompressed(archive) / info->size;
+		    			percent = (double)archive_position_uncompressed(archive) / info->size;
 				if (needdisp == 0) {
 					PROGRESS(trans, cb_state, what, (int)(percent * 100), howmany, (howmany - remain +1));
 				}
